requirements.txt
Flask
flask-cors
tensorflow
numpy
pillow
opencv-python-headless
gunicorn